requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
flask
flask-cors
opencv-python
numpy
pillow